action_anticipation_module/EGO4D-prediction/requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
av==10.0.0
bitsandbytes==0.37.2
black==21.4b2
cachetools==5.3.0
click==8.1.3
clip==1.0
cloudpickle==2.2.1
decord==0.6.0
editdistance==0.6.0
einops==0.3.0
fonttools==4.38.0
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.1
huggingface-hub==0.14.1
hydra-core==1.3.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
imutils==0.5.4
iopath==0.1.9
lmdb==1.2.1
markupsafe==2.1.2
matplotlib==3.5.3
multidict==6.0.4
mypy-extensions==1.0.0
networkx==2.6.3
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
openapi-schema-pydantic==1.2.4
opencv-python==4.5.3.56
pandas==1.3.5
parameterized==0.8.1
pathspec==0.11.0
pillow==9.4.0
portalocker==2.7.0
protobuf==3.20.3
pyasn1-modules==0.2.8
pycocotools==2.0.6
pydantic==1.10.7
pydeprecate==0.3.1
pydot==1.4.2
pytorch-lightning==1.5.6
pytorchvideo==0.1.5
regex==2022.10.31
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.24.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==59.5.0
sqlalchemy==2.0.12
submitit==1.3.3
tabulate==0.9.0
tensorboard==2.11.2
termcolor==2.2.0
torch==1.9.0
torchmetrics==0.11.1
torchvision==0.10.0
typed-ast==1.5.4
yacs==0.1.8
yarl==1.8.2
zipp==3.12.1